pyproject.toml
[project]
name = "earthquakes"
version = "0.1.0"
description = ""
authors = [
    {name = "bntmzu",email = "kuranova988@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp (>=3.11.16,<4.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
